~ sorting project boilerplate, license, etc.

+++ src/win.volumes.c
@@ -1,3 +1,49 @@
+/* /////////////////////////////////////////////////////////////////////////
+ * File:    win.volumes.c
+ *
+ * Purpose: Synesis Software Disk Utility library, for Windows: Volumes API.
+ *
+ * Created: 4th April 2019
+ * Updated: 2nd August 2019
+ *
+ * Home:    http://github.com/synesissoftware/ss-win-diskutil
+ *
+ * Copyright (c) 2019, Matthew Wilson and Synesis Software
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are
+ * met:
+ *
+ * - Redistributions of source code must retain the above copyright notice,
+ *   this list of conditions and the following disclaimer.
+ * - Redistributions in binary form must reproduce the above copyright
+ *   notice, this list of conditions and the following disclaimer in the
+ *   documentation and/or other materials provided with the distribution.
+ * - Neither the name(s) of Matthew Wilson and Synesis Software nor the
+ *   names of any contributors may be used to endorse or promote products
+ *   derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+ * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ * ////////////////////////////////////////////////////////////////////// */
+
+
+/* /////////////////////////////////////////////////////////////////////////
+ * includes
+ */
+
+#include <stdint.h>
 
 #include <ss-win-diskutil/volumes.h>
 
@@ -506,8 +552,8 @@ internal_append_(
                 vol->id.ptr             +=  d / sizeof(wchar_t);
                 vol->friendlyName.ptr   +=  d / sizeof(wchar_t);
 
-                _ASSERTE(vol->id.len == lstrlenW(vol->id.ptr));
-                _ASSERTE(vol->friendlyName.len == lstrlenW(vol->friendlyName.ptr));
+                _ASSERTE(vol->id.len == (size_t)lstrlenW(vol->id.ptr));
+                _ASSERTE(vol->friendlyName.len == (size_t)lstrlenW(vol->friendlyName.ptr));
             }
         }
